requirements.txt
moderngl==5.12.0
numpy==2.3.5
pillow==12.0.0
scipy==1.16.3
textual==6.6.0